requirements.txt
numpy==1.11.2
biopython==1.68